% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntbSimViz.R
\name{ntbSimViz}
\alias{ntbSimViz}
\title{Neutral Theory of biodiversity simulation and visualization}
\usage{
ntbSimViz(fname, J, m, nu, Sm, SADm, nstep, init = NULL)
}
\arguments{
\item{fname}{file name to save the animation to, ending in .mp4}

\item{J}{size of the local community (i.e. number of individuals)}

\item{m}{migration probability}

\item{nu}{speciation probability}

\item{Sm}{number of species in the metacommunity}

\item{SADm}{function defining the metacommunity SAD as a \code{pika::sad}
class object}

\item{nstep}{number of time steps to simulate}

\item{init}{initial state (optional, defults to NULL)}
}
\value{
Saves an animation and silently returns the final state of the
community as a list with elements
}
\description{
Simulate communities under the neutral theory of biodiversity
}
\examples{
x <- 'foo'

}
